pyproject.toml
[tool.poetry]
name = "toucan"
version = "0.1.0"
description = "FastAPI backend with Supabase auth"
authors = ["Drew <drew@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
python-dotenv = "^1.0.0"
supabase = "^1.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 